pyproject.toml
[project]
name = "mini-job-platform-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
]
